Exp list
Dodany exp list w panelu lekarza, drobne zmiany w rejestracji pacjenta.
