+++ src/app/admin/add-user/add-user.component.scss
@@ -4,7 +4,7 @@ div.wrapper {
     justify-content: center;
     align-items: center;
     width: 100%;
-    height: 100%;
+    
     div.inner-wrapper {
         width: 100%;
         height: 100%;
