+++ src/app/lekarz/lekarz.module.ts
@@ -6,11 +6,13 @@ import { BadanieFizykalneComponent } from './badanie-fizykalne/badanie-fizykalne
 import { LekarzRoutingModule } from './lekarz-routing.module';
 import { MaterialModule } from '../material/material.module';
 import { BadanieLaboratoryjneComponent } from './badanie-laboratoryjne/badanie-laboratoryjne.component';
-
+import { PreviousVisitsComponent } from './wizyta/previous-visits/previous-visits.component';
+import { ListExaminationsComponent } from './wizyta/list-examinations/list-examinations.component';
+import { ListLabExaminationsComponent } from './wizyta/list-lab-examinations/list-lab-examinations.component';
 
 
 @NgModule({
-  declarations: [ListaWizytComponent, WizytaComponent, BadanieFizykalneComponent, BadanieLaboratoryjneComponent],
+  declarations: [ListaWizytComponent, WizytaComponent, BadanieFizykalneComponent, BadanieLaboratoryjneComponent, PreviousVisitsComponent, ListExaminationsComponent, ListLabExaminationsComponent],
   imports: [
     LekarzRoutingModule,
     CommonModule,
