+++ src/app/lekarz/wizyta/wizyta.component.scss
@@ -16,7 +16,7 @@ div.wrapper {
         }
     }
     div.inner-wrapper {
-        width: 90%;
+        width: 80%;
         height: 100%;
         
         background: #fff;
@@ -35,6 +35,8 @@ div.wrapper {
                 textarea {
                     margin-top: 25px;
                     min-height: 50px;
+                    //background-color: #f3f3f3;
+                    border-radius: 3px;
                 }
             }
             div.double-input {
