+++ src/app/lekarz/wizyta/wizyta.component.ts
@@ -4,30 +4,7 @@ import { MatTableDataSource, MatSort, MatPaginator, MatDialog } from '@angular/m
 import { BadanieFizykalneComponent } from '../badanie-fizykalne/badanie-fizykalne.component';
 import { BadanieLaboratoryjneComponent } from '../badanie-laboratoryjne/badanie-laboratoryjne.component';
 
-export interface Visit {
-  id: number;
-  docName: string;
-  desc: string;
-  diagnose: string;
-  date: any;
-}
 
-export interface LabExamination {
-  id: number;
-  docName: string;
-  workerName: any;
-  type: string;
-  result: string;
-  date: any;
-}
-
-export interface Examination {
-  id: number;
-  docName: string;
-  type: string;
-  result: string;
-  date: any;
-}
 
 @Component({
   selector: 'app-wizyta',
@@ -67,54 +44,17 @@ export class WizytaComponent implements OnInit {
     });
   }
 
-  public Visits: Visit[] = [];
-  public LabExaminations: LabExamination[] = [];
-  public Examinations: Examination[] = [];
-
-  visitsDisplayedColumns: string[] = ['position', 'doc_name', "description", 'diagnose', 'date'];
-  labExaminationsDisplayedColumns: string[] = ['position', 'doc_name', 'worker_name', 'type' , 'result', 'date'];
-  examinationsDisplayedColumns: string[] = ['position', 'doc_name', 'type', 'result', 'date'];
-  visits = new MatTableDataSource(this.Visits);
-  labExaminations = new MatTableDataSource(this.LabExaminations);
-  examinations = new MatTableDataSource(this.Examinations);
-
-
-  @ViewChild(MatSort, {static: true}) sort: MatSort;
-  @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
-  
-  applyFilter(filterValue: string) {
-    this.visits.filter = filterValue.trim().toLowerCase();
-    this.labExaminations.filter = filterValue.trim().toLowerCase();
-    this.examinations.filter = filterValue.trim().toLowerCase();
-  }
-
   private data;
 
   private handleParams(a){
     console.log(a.id);
     this.data = a.id;
   }
-
+  
   ngOnInit() {
     this.route.params.subscribe({
       next: this.handleParams.bind(this)
     });
-    for(let i = 0; i < 696; i++) {
-      this.Visits.push({ id: i, desc: "Potrzebna decyzja czy pracownik laboratorium może widzieć wykonane badania innego pracownika", diagnose: "rak", date: '11/13/2019', docName: "Czesiek"});
-    }
-    for(let i = 0; i < 2; i++) {
-      this.LabExaminations.push({ id: i, docName: "Czesiek", workerName: "Macin", type: "Badanie wkurwa", result: "Oh yes" , date: '21/37/2019', });
-    }
-    for(let i = 0; i < 3; i++) {
-      this.Examinations.push({ id: i, docName: "Macin", type: "Badanie dupy", result: "OH YES" , date: '13/37/2019', });
-    }
-
-    this.visits.paginator = this.paginator;
-    this.visits.sort = this.sort;
-    this.labExaminations.paginator = this.paginator;
-    this.labExaminations.sort = this.sort;
-    this.examinations.paginator = this.paginator;
-    this.examinations.sort = this.sort;
   }
