+++ src/app/recepcja/dodaj-pacjenta/dodaj-pacjenta.component.html
@@ -1,6 +1,6 @@
 <div class="wrapper">
     <div class="inner-wrapper">
-        <form (submit)="onSubmit($event)" ngNativeValidate> 
+        <form (submit)="onSubmit($event)"> 
             <h2>Dodawanie nowego pacjenta do bazy</h2>
             <div class="double-input">
                 <mat-form-field>
